fix(sync): isolate entity recovery ownership
